project_management_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
passlib==1.7.4
python-jose==3.3.0
pydantic[email]==1.10.2
redis==4.3.3
bcrypt==3.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
celery==5.2.3
requests==2.26.0
